requirements.txt
django-allauth
psycopg2-binary
django-environ
Django
gunicorn